Factor out toString-truncation
