+++ src/main/scala/org/hammerlab/guacamole/loci/set/Contig.scala
@@ -3,8 +3,8 @@ package org.hammerlab.guacamole.loci.set
 import java.lang.{Long => JLong}
 
 import com.google.common.collect.{RangeSet, TreeRangeSet, Range => JRange}
-import org.hammerlab.guacamole.Common
 import org.hammerlab.guacamole.loci.SimpleRange
+import org.hammerlab.guacamole.strings.TruncatedToString
 
 import scala.collection.JavaConversions._
 import scala.collection.mutable.ArrayBuffer
@@ -12,7 +12,7 @@ import scala.collection.mutable.ArrayBuffer
 /**
   * A set of loci on a contig, stored/manipulated as loci ranges.
   */
-case class Contig(name: String, private val rangeSet: RangeSet[JLong]) {
+case class Contig(name: String, private val rangeSet: RangeSet[JLong]) extends TruncatedToString {
 
   /** Is the given locus contained in this set? */
   def contains(locus: Long): Boolean = rangeSet.contains(locus)
@@ -67,20 +67,11 @@ case class Contig(name: String, private val rangeSet: RangeSet[JLong]) {
     (Contig(name, firstRanges), Contig(name, secondRanges))
   }
 
-  override def toString: String = truncatedString(Int.MaxValue)
-
-  /** String representation, truncated to maxLength characters. */
-  private def truncatedString(maxLength: Int = 100): String = {
-    Common.assembleTruncatedString(stringPieces(), maxLength)
-  }
-
   /**
-   * Iterator over string representations of each range in the map.
-   *
-   * If includeValues is true (default), then also include the values mapped to by this LociMap. If it's false,
-   * then only the keys are included.
-   */
-  private[set] def stringPieces() = {
+    * Iterator over string representations of each range in the map, used to assemble (possibly truncated) .toString()
+    * output.
+    */
+  def stringPieces = {
     ranges.iterator.map(pair =>
       "%s:%d-%d".format(name, pair.start, pair.end)
     )
